requirements.txt
Flask==2.0.1
auth0-python==3.15.0
python-dotenv==0.19.1
